logger: Add a Level() method b/c we sometimes need different behavior depending on level
