+++ internal/synclet/server_adapter.go
@@ -1,14 +1,11 @@
 package synclet
 
 import (
-	"context"
 	"fmt"
 	"sync"
 
 	"github.com/docker/distribution/reference"
 	"github.com/windmilleng/tilt/internal/k8s"
-	"github.com/windmilleng/tilt/internal/logger"
-	"github.com/windmilleng/tilt/internal/output"
 
 	"github.com/windmilleng/tilt/internal/model"
 	"github.com/windmilleng/tilt/internal/synclet/proto"
@@ -23,37 +20,6 @@ func NewGRPCServer(del *Synclet) *GRPCServer {
 }
 
 var _ proto.SyncletServer = &GRPCServer{}
-
-func logLevelToProto(level logger.Level) (proto.LogLevel, error) {
-	switch level {
-	case logger.InfoLvl:
-		return proto.LogLevel_INFO, nil
-	case logger.VerboseLvl:
-		return proto.LogLevel_VERBOSE, nil
-	case logger.DebugLvl:
-		return proto.LogLevel_DEBUG, nil
-	default:
-		return proto.LogLevel_INFO, fmt.Errorf("unknown log level '%v'", level)
-	}
-}
-
-func makeContext(ctx context.Context, logStyle *proto.LogStyle, f func(m *proto.LogMessage) error) (context.Context, error) {
-	writeLog := func(level logger.Level, bytes []byte) error {
-		protoLevel, err := logLevelToProto(level)
-		if err != nil {
-			return err
-		}
-
-		logMessage := &proto.LogMessage{Level: protoLevel, Message: bytes}
-		return f(logMessage)
-	}
-
-	l := logger.NewFuncLogger(logStyle.ColorsEnabled, writeLog)
-
-	// TODO(matt) making a new outputter here is hacky - since outputter is stateful, someone might make
-	// rely on state persisting across service boundaries
-	return output.WithOutputter(logger.WithLogger(ctx, l), output.NewOutputter(l)), nil
-}
 
 func (s *GRPCServer) GetContainerIdForPod(req *proto.GetContainerIdForPodRequest, server proto.Synclet_GetContainerIdForPodServer) error {
 	sendMutex := new(sync.Mutex)
